refactor(category): Category 및 CategoryType Repository 의존성 역전 적용
- CategoryRepository: Application Port로 JPA 의존성 제거
- CategoryTypeRepository: Application Port로 JPA 의존성 제거
- Infrastructure에 각각 Adapter 구현체와 JpaRepository 분리
- ArchUnit 테스트 통과를 위한 레이어 의존성 개선

+++ src/main/java/com/myfinance/ledger/interfaces/rest/category/CategoryTypeController.java
@@ -45,7 +45,7 @@ public class CategoryTypeController {
      * DELETE /api/category-types/{id}
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategoryType(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteCategoryType(@PathVariable Long id) {
         categoryTypeService.deleteCategoryType(id);
         return ResponseEntity.noContent().build();
     }
